backend/requirements.txt
# Core
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1

# LangChain / LangGraph
langgraph==0.4.1
langchain==0.3.25
langchain-anthropic==0.3.15
langchain-community==0.3.21

# Embedding
sentence-transformers==3.4.1

# Supabase
supabase==2.13.0

# Web scraping / document extraction
trafilatura==2.0.0
pymupdf==1.25.3

# Search
tavily-python==0.5.0

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
pytest-mock==3.14.0
httpx==0.28.1
